core: split read cache internals
